+++ database/migrations/2016_09_14_144841_messages_table.php
@@ -18,7 +18,7 @@ class MessagesTable extends Migration {
 			$table->string("text");
 			$table->integer("estado");
 			$table->integer("conversation_id");
-			$table->bigInteger("user_id")->unsigned();
+			$table->integer("user_id")->unsigned();
 			$table->foreign('user_id')->references('id')->on('users')
 			                             ->onDelete('cascade')
 										 ->onUpdate('cascade');
